+++ packages/ranger-ui/src/ElfsightWidget/index.ts
@@ -1,0 +1,2 @@
+export type { ElfsightWidgetProps } from './elfsightWidget'
+export { default as ElfsightWidget } from './elfsightWidget'