+++ apps/ranger-vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ranger-theme/vite
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/ui@0.9.0
+
 ## 0.1.17
 
 ### Patch Changes